HW15/requirements.txt
pytest == 7.1.2
selenium == 4.2.0
webdriver-manager == 3.7.0
requests==2.28.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
